Solution that actually works this time
Former-commit-id: eedf621bc5fe7a10d42f5186525853a8075c920d

+++ OpenGL_Template/Buffer.cpp
@@ -96,7 +96,7 @@ std::uint32_t Buffer::Append(std::size_t size, void* data) {
 	std::size_t offset = m_size;
 	Resize(offset + size);
 	Upload(offset, size, data);
-	return offset;
+	return (uint32_t)offset;
 }
 
 // // RESIZE
